Add support for multiple input fastqs
Sequencing a single cell library across multiple lanes is common practice and cite-seq-count should
be able to handle that and not expect the user to merge them downstream. To reduce the amount of
new code added to handle this, the current preprocessing is run on each input file and then merged
together before barcode correction.